remove unused method and rename
