+++ Application/EdFi.Ods.AdminApp.Management/Database/Commands/ValidationConstants.cs
@@ -3,17 +3,10 @@
 // The Ed-Fi Alliance licenses this file to you under the Apache License, Version 2.0.
 // See the LICENSE and NOTICES files in the project root for more information.
 
-using EdFi.Admin.DataAccess.Models;
-
-namespace EdFi.Ods.AdminApp.Management.Database.Queries
+namespace EdFi.Ods.AdminApp.Management.Database.Commands
 {
-    public static class ApplicationExtensions
+    public static class ValidationConstants
     {
-        public static bool IsSystemReservedApplication(this Application application)
-        {
-            return application != null && application.Vendor.IsSystemReservedVendor();
-        }
-
         public static int MaximumApplicationNameLength = 50;
     }
 }
